backend/requirements.txt
fastapi>=0.111.0
uvicorn>=0.29.0
python-multipart>=0.0.9
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4,<2.0.0
python-dotenv>=1.0.1
httpx>=0.27.0
pydantic>=2.7.0,<3.0.0
pinecone>=3.2.2
sentence-transformers>=2.7.0
langchain>=0.1.0
langchain-community>=0.0.38
langchain-core>=0.3.65,<1.0.0
langchain-text-splitters>=0.0.2
langchain-huggingface>=0.3.0
packaging>=24.2,<25.0
